requirements.txt
python-dotenv
pydantic
pydantic_core